feat: add jitter to retry backoff

+++ migrations/20230117003150_add_retry_columns_to_issues_queue.sql
@@ -1,5 +1,5 @@
 ALTER TABLE issue_delivery_queue
-ADD COLUMN n_retries int NOT NULL DEFAULT 0;
+ADD COLUMN n_retries smallint NOT NULL DEFAULT 0;
 
 ALTER TABLE issue_delivery_queue
 ADD COLUMN execute_after timestamptz NOT NULL DEFAULT NOW();
